Finalize rename
