--- CNUSPACKER FIX/CNUSPACKER.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
-    <RootNamespace>CNUSPACKER_FIX</RootNamespace>
-  </PropertyGroup>
-
-</Project>
