+++ CNUSPACKER.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CNUSPACKER", "CNUSPACKER FIX\CNUSPACKER.csproj", "{DC99F744-E0A0-41F2-A1CA-85B28AB1FE4B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CNUSPACKER", "CNUSPACKER\CNUSPACKER.csproj", "{DC99F744-E0A0-41F2-A1CA-85B28AB1FE4B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
